requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
google-cloud-bigquery>=3.0.0
db-dtypes>=1.0.0
numpy>=1.24.0
